sampleFPT
