+++ scenes/classroom/scene.json
