+++ scenes/disney-bsdf/scene.json
