+++ scenes/teapot/scene.json
